third_party/blink/web_tests/fast/scrolling/subpixel-overflow-mouse-drag-expected.txt

PASS container.scrollTop is within 0.02 of 0.57
PASS successfullyParsed is true

TEST COMPLETE

Tests scroll behavior when a subpixel scrollbar thumb is dragged by mouse.
